requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
python-dotenv==1.0.0
shortuuid==1.0.11
python-dateutil==2.8.2
requests==2.31.0
paypal-checkout-serversdk==1.0.1
Werkzeug==2.3.7
